+++ src/Errors.tsx
@@ -6,6 +6,9 @@ export type ErrorCode =
     "ERR_SET_INBRAIN_VALUES" |
     "ERR_GET_REWARDS" |
     "ERR_CONFIRM_REWARDS" |
+    "ERR_CHECK_SURVEYS_AVAILABLE" |
+    "ERR_GET_NATIVE_SURVEYS" |
+    "ERR_SHOW_NATIVE_SURVEY" |
     "ERR_SHOW_SURVEYS" |
     "ERR_SET_LANGUAGE" |
     "ERR_SET_TITLE" |
@@ -20,6 +23,9 @@ export type ErrorCode =
     ERR_SET_INBRAIN_VALUES: "Error while initialising sdk",
     ERR_GET_REWARDS: "Error while getting rewards",
     ERR_CONFIRM_REWARDS: "Error while confirming rewards",
+    ERR_CHECK_SURVEYS_AVAILABLE: "Error while checking surveys availability",
+    ERR_GET_NATIVE_SURVEYS: "Error while getting native surveys",
+    ERR_SHOW_NATIVE_SURVEY: "Error while showing native survey",
     ERR_SHOW_SURVEYS: "Error while showing surveys",
     ERR_SET_TITLE: "Error while setting title",
     ERR_SET_LANGUAGE: "Error while setting language",